Move the neutral schema modules into @modelcontextprotocol/core
The MCP spec schemas (types/schemas.ts), the OAuth/OpenID schemas
(shared/auth.ts), the protocol constants (types/constants.ts), and the three
JSON value types move verbatim from core-internal into packages/core/src —
core now owns the schema sources instead of bundling them out of
core-internal through build-only subpath aliases.

Mechanism:
- core's curated root entry re-exports the same public surface as before
  (unchanged 172 exports), now from its own local modules.
- A new ./internal subpath re-exports the moved modules wholesale for the
  sibling SDK packages (internal helper schemas, auth types, constants, JSON
  types — names that are deliberately not public on the root entry).
- core-internal keeps the old module paths as one-to-one named re-export
  shims, so no importer or test changes anywhere.
- client/server/server-legacy declare @modelcontextprotocol/core as a real
  dependency; their bundles keep @modelcontextprotocol/core/internal as an
  external runtime import (explicit tsdown external entry) instead of
  carrying their own bundled schema copies.
- The build-only core-internal/schemas + core-internal/auth tsconfig/tsdown
  aliases are deleted everywhere; the per-package tsconfigs gain a single
  source-first alias for @modelcontextprotocol/core/internal so typecheck
  and vitest stay build-order independent.

Layering stays acyclic: core depends only on zod; core-internal depends on
core; the wire era modules are untouched (their frozen copies and their
runtime constants imports resolve through the shims unchanged).
