+++ packages/client/tsdown.config.ts
@@ -34,5 +34,8 @@ export default defineConfig({
         }
     },
     noExternal: ['@modelcontextprotocol/core-internal', 'ajv', 'ajv-formats', '@cfworker/json-schema'],
-    external: ['@modelcontextprotocol/client/_shims']
+    // The schema modules live in @modelcontextprotocol/core (a real runtime dependency); the
+    // bundled core-internal shims import them via the './internal' subpath, which must stay an
+    // external import (explicit entry — the tsconfig paths alias would otherwise inline it).
+    external: ['@modelcontextprotocol/client/_shims', '@modelcontextprotocol/core/internal']
 });
