+++ packages/core/src/index.ts
@@ -2,8 +2,9 @@
 //
 // Canonical public home for the Model Context Protocol specification + OAuth/OpenID Zod schemas.
 //
-// These are the exact schema constants the SDK validates against internally (defined in the
-// private @modelcontextprotocol/core-internal package). This package bundles core-internal and re-exports ONLY the
+// These are the exact schema constants the SDK validates against internally — the schema source
+// modules live in THIS package (src/schemas.ts, src/auth.ts) and the private
+// @modelcontextprotocol/core-internal package re-exports them. This root entry re-exports ONLY the
 // `*Schema` Zod values, so consumers can validate protocol/OAuth payloads directly — e.g.
 // `CallToolResultSchema.parse(value)` / `.safeParse(value)` — without depending on core-internal's
 // barrel.
@@ -11,13 +12,12 @@
 // Scope: Zod schemas ONLY. The corresponding spec TypeScript types, error classes, enums, and
 // type guards are part of the public API of @modelcontextprotocol/server and /client.
 //
-// Two groups, kept separate to mirror core-internal's own spec-vs-auth split, each bundled from a build-only
-// subpath alias of core-internal (tsconfig.json + tsdown.config.ts):
-//   - SPEC schemas, from @modelcontextprotocol/core-internal/schemas (core-internal/src/types/schemas.ts): every
-//     `export const *Schema` EXCEPT internal helpers with no public spec type (e.g.
-//     BaseRequestParamsSchema). Mirrors core-internal's SPEC_SCHEMA_KEYS allowlist.
-//   - OAUTH/OPENID schemas, from @modelcontextprotocol/core-internal/auth (core-internal/src/shared/auth.ts).
-// The coreSchemas test asserts both groups stay in sync with their core-internal source modules.
+// Two groups, kept separate to mirror the SDK's own spec-vs-auth split:
+//   - SPEC schemas, from ./schemas (src/schemas.ts): every `export const *Schema` EXCEPT internal
+//     helpers with no public spec type (e.g. BaseRequestParamsSchema). Mirrors core-internal's
+//     SPEC_SCHEMA_KEYS allowlist.
+//   - OAUTH/OPENID schemas, from ./auth (src/auth.ts).
+// The coreSchemas test asserts both groups stay in sync with their source modules.
 export {
     AnnotationsSchema,
     AudioContentSchema,
@@ -179,10 +179,10 @@ export {
     UnsubscribeRequestSchema,
     UntitledMultiSelectEnumSchemaSchema,
     UntitledSingleSelectEnumSchemaSchema
-} from '@modelcontextprotocol/core-internal/schemas';
+} from './schemas';
 
 // Auth schemas (OAuth / OpenID / IdJag) — kept as a SEPARATE group from the MCP spec schemas above,
-// mirroring core-internal's own spec-vs-auth split (these live in core-internal/src/shared/auth.ts, not types/schemas.ts,
+// mirroring the SDK's own spec-vs-auth split (these live in src/auth.ts, not src/schemas.ts,
 // and are registered as `authSchemas` in core-internal's specTypeSchema.ts). This group is EXACTLY core-internal's
 // `authSchemas` set — every auth schema that has a public spec type (so `isSpecType.OAuthTokens`,
 // `isSpecType.IdJagTokenExchangeResponse`, etc. exist). The typeless internal URL field-validators
@@ -201,4 +201,4 @@ export {
     OAuthTokensSchema,
     OpenIdProviderDiscoveryMetadataSchema,
     OpenIdProviderMetadataSchema
-} from '@modelcontextprotocol/core-internal/auth';
+} from './auth';
