+++ packages/core/src/types.ts
@@ -1,0 +1,4 @@
+/* JSON types — the base value vocabulary the spec schemas are typed against. */
+export type JSONValue = string | number | boolean | null | JSONObject | JSONArray;
+export type JSONObject = { [key: string]: JSONValue };
+export type JSONArray = JSONValue[];
