+++ packages/server-legacy/package.json
@@ -83,6 +83,7 @@
         "test:watch": "vitest"
     },
     "dependencies": {
+        "@modelcontextprotocol/core": "workspace:^",
         "zod": "catalog:runtimeShared",
         "raw-body": "catalog:runtimeServerOnly",
         "content-type": "catalog:runtimeShared",
